+++ openapi/oh.yaml
@@ -6928,9 +6928,9 @@ components:
           type: boolean
         male:
           type: boolean
-        pharmacy:
-          type: boolean
         opd:
+          type: boolean
+        pharmacy:
           type: boolean
       required:
       - beds
@@ -7759,15 +7759,6 @@ components:
           description: "Assessment performed before the patient's admission. Example:\
             \ preliminary examinations, lab results, medical observations."
           example: Blood test and X-ray prior to admission
-        conditionAtAdmission:
-          type: array
-          description: "Patient's condition at the time of admission. Stored as a\
-            \ list of descriptors (e.g., stable, critical, unconscious)."
-          example:
-          - stable
-          - conscious
-          items:
-            type: string
         userID:
           type: string
           description: User id
@@ -8673,6 +8664,19 @@ components:
           type: string
           description: Conditioning others
           example: others
+        conditionAtAdmission:
+          type: array
+          description: "Patient's condition at the time of admission. Stored as a\
+            \ list of descriptors (e.g., stable, critical, unconscious)."
+          example:
+          - stable
+          - conscious
+          items:
+            type: string
+        tdr:
+          type: string
+          description: 'Conditioning tdr '
+          example: sonde
         patient:
           $ref: "#/components/schemas/PatientDTO"
           description: patient
